requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
pillow
tensorflow
python-multipart
